moved JointLimitRange
